ml_service/requirements.txt
Flask==1.0.2
tensorflow==2.1.0
efficientnet==1.0.0
numpy==1.16.2
Pillow==6.2.0
